requirements.txt
numpy
tensorflow-gpu
pandas
ipykernel
pillow
matplotlib
tensorflow-datasets
scipy